requirements.txt
python_dotenv
langchain
langchain_openai
langchain_community
langchain_core
langgraph
pillow
pymongo

langgraph-checkpoint-mongodb
